Warning, paredes, controles, arreglos, optimización
Enemigos, colliders, ....

+++ Vamoos2/Assets/Scripts/Enemigos/Enemigos.cs
@@ -12,7 +12,7 @@ public class Enemigos : MonoBehaviour
     protected float maxHealth;
     protected float currentHealth;
     protected Vector3 mov;
-    protected int armor = 5;
+    protected float armor = 5;
     protected Animator animE;
 
     protected GameObject shield;
